Declare cookies as secure if on HTTPS page
During one of our monthly security changes we noticed that this cookie does not include the "secure" flag. It's not a biggie at all but  we had to make this change in our ProcessWire core installation to prevent the security tool from reporting it. 

Can we merge this into the core so we can keep our code base upgradable?