Add support for $urls(__FILE__);
This simple addition makes it a lot easier to convert file paths to urls relative to the PW root. It's not good to do this via `str_replace($config->paths->root, '/', __FILE__)` because this does not work if $config->urls->root is different to '/' (subfolder installations). Doing `str_replace($config->paths->root, $config->urls->root, __FILE__)` is also not a proper solution because it does not work with windows filepaths!

IMHO there needs to be a quick, simple and robust solution for that and with this little fix to $config->urls/paths->get() we get one:

![img](https://i.imgur.com/lVG6iIg.png)