User feedback enhancements to ProcessForgotPassword.module
Refactored to allow calling processInput() on the initial password reset request so that the form will show an error when the username or email field is blank. Also added user feedback when beHonest is true and an invalid email/username is entered, instead of redirecting the user with no message. Made the display of the Continue button upon display of InlineNotices optional, since this is redundant when used in conjunction with the Login/Register module, which has its own back button.