New option: Remember Language Tabs Focus
This new option gives editors a better workflow when working with languages and the Language Tab Module. Until now, the focus of the current language was lost after each save and the editor always had to switch back to the language they were currently editing. When editing a language version for a longer period of time, the new option now improves the workflow. By double-clicking on the desired language version, the status is saved in a cookie for the current session if the Remember option is activated. After each save, the focus is now automatically set to the last selected language. See original ticket: processwire/processwire-issues#1622