MarkupRSS: Allow markup in <description>.
This removes entity encoding from the <description> element and adds the conversion of relative to absolute URLs that was previously only applied to <content>. It also extends this conversion to account for single quotes such as href='/path'.

According to the spec you’re supposed to put the entire content into <description> if you’re not using summaries. This change makes that feasible.