Update Wire.php: Allow registering new notice types
For a project of mine I wanted to use a `NoticeSuccess` type to show success-alerts on the frontend. This works as I would expect (create a child class of Notice and return 'success' in the getName() method), but a notice is throwing at this line, that `$this->_notices['success']` is undefined.

My proposed change would not alter the functionality but would get rid of the notice, as `empty()`, just like `isset()`, does not throw a warning when the key is undefined.